Refactored base_GFormula.py & GFormula.py and moved to "time_varying" module.
